boy. This commit was a long time coming.
I cut my losses on the results bit. Wrote in a check for 2018 & in the case of 2018 I manually entered in results for TWFighter. I intend to come back to this and find an elegant solution- Most likely not for this project, but for personal understanding. Too frustrated with the problem to give it up